[fixed] 팔로우 하지 않은 사용자일 경우 deleteFollow가 실행되지 않도록 수정

+++ src/users/users.service.ts
@@ -144,6 +144,17 @@ export class UsersService {
     async deleteFollow(followerId: number, followeeId: number, qr?: QueryRunner) {
         const repository = this.getUserFollowersRepository(qr);
 
+        const existingRelation = await repository.findOneBy({
+            follower: { id: followerId },
+            followee: { id: followeeId },
+        });
+
+        if (!existingRelation) {
+            throw new BadRequestException(
+                `팔로우 하지 않은 사용자입니다.`
+            );
+        }
+
         await repository.delete({
             follower: {
                 id: followerId,
